# public/res/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "An api for interacting with an E-Commerce server.\n\n### Admin user info:\n##### Username | 'admin'\n##### Password | 'password'",
    "version": "1.0.0",
    "title": "Online E-Commerce API",
    "contact": {
      "email": "arigorn15@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://localhost:3000/",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/users": {
      "summary": "The user's information",
      "description": "Edit user's information",
      "put": {
        "summary": "Edit user's information",
        "description": "Edit information for a specific user, if that user is the logged in user and return the changed user information.",
        "operationId": "put_user",
        "requestBody": {
          "description": "User Information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user_update"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "The id of the user.",
            "required": false,
            "schema": {
              "type": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                  "username": "test_user",
                  "email": "john_doe@email.com",
                  "first_name": "John,",
                  "last_name": "Doe",
                  "role": "user"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "User with ID {userId} not found"
                }
              }
            }
          }
        },
        "tags": ["User"]
      }
    },
    "/users/password": {
      "summary": "User Password Operations",
      "description": "Change user's password",
      "put": {
        "summary": "Change user's password",
        "description": "Change password for a specific user, if that user is the logged in user and return the changed user information.",
        "operationId": "put_user_password",
        "requestBody": {
          "description": "User Password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user_password_update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": {
                  "message": "Password changed successfully!"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in or incorrect password",
            "content": {
              "application/json": {
                "example1": {
                  "status": 401,
                  "message": "Please Log In!"
                },
                "example2": {
                  "status": 401,
                  "message": "Incorrect Password"
                }
              }
            }
          },
          "404": {
            "description": "Not logged in or incorrect password",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "User with ID <userId> not found"
                }
              }
            }
          }
        },
        "tags": ["User"]
      }
    },
    "/users/admin": {
      "summary": "User Admin Operations",
      "description": "Get users",
      "get": {
        "summary": "Get a list of users",
        "description": "Administrator can get a list of users filtered by email, username, or role",
        "operationId": "get_users",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "filter response by email, username, or role",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "username": "test_user",
                    "email": "john_doe@email.com",
                    "first_name": "John,",
                    "last_name": "Doe",
                    "role": "user"
                  },
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "username": "test_user2",
                    "email": "jane_doe@email.com",
                    "first_name": "Jane,",
                    "last_name": "Doe",
                    "role": "admin"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No User session exists",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "403": {
            "description": "Not administrator",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          },
          "404": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "Users not found!"
                }
              }
            }
          }
        },
        "tags": ["User"]
      }
    },
    "/users/admin/{userId}": {
      "summary": "User Admin Operations",
      "description": "Edit a User's information",
      "patch": {
        "summary": "Edit user's information",
        "description": "Administrator can edit the email address and role for a specific user, return the list of users with updated information.",
        "operationId": "patch_user",
        "requestBody": {
          "description": "Updated User email and role",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user_patch"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user.",
            "required": true,
            "schema": {
              "type": "uuid"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "filter response by email or username",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "username": "test_user",
                    "email": "john_doe@email.com",
                    "first_name": "John,",
                    "last_name": "Doe",
                    "role": "user"
                  },
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "username": "test_user2",
                    "email": "jane_doe@email.com",
                    "first_name": "Jane,",
                    "last_name": "Doe",
                    "role": "admin"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Malformed Request",
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "message": "Invalid Input!"
                }
              }
            }
          },
          "401": {
            "description": "No User session exists",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "403": {
            "description": "Not administrator",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          },
          "404": {
            "description": "User not found or invalid filter",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "User with ID <userId> not found"
                },
                "example2": {
                  "status": 404,
                  "message": "Users not found!"
                }
              }
            }
          }
        },
        "tags": ["User"]
      },
      "delete": {
        "summary": "Delete a user",
        "description": "Administrator can delete a specific user, return the list of users with updated information.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user.",
            "required": true,
            "schema": {
              "type": "uuid"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "filter response by email or username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "username": "test_user",
                    "email": "john_doe@email.com",
                    "first_name": "John,",
                    "last_name": "Doe",
                    "role": "user"
                  },
                  {
                    "id": "00a725ab-9638-4a4b-b2da-449ec0201b07",
                    "username": "test_user2",
                    "email": "jane_doe@email.com",
                    "first_name": "Jane,",
                    "last_name": "Doe",
                    "role": "admin"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No User session exists",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "message": "Please Log In!"
                }
              }
            }
          },
          "403": {
            "description": "Not administrator",
            "content": {
              "application/json": {
                "example": {
                  "status": 403,
                  "message": "Permission Denied!"
                }
              }
            }
          },
          "404": {
            "description": "User not found or invalid filter",
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "message": "User with ID <userId> not found"
                },
                "example2": {
                  "status": 404,
                  "message": "Users not found!"
                }
              }
            }
          }
        },
        "tags": ["User"]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "user_update": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User_Registration"
        }
      },
      "user_patch": {
        "type": "object",
        "properties": {
          "email": "string",
          "role": "string"
        }
      },
      "user_password_update": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User_Password_Update"
        }
      },
...
    }
  }
}